finalize inputbox severity
